requirements.txt
openai==2.17.0
google-genai==1.62.0
google-generativeai==0.6.15
python-dotenv==1.2.1
numpy==2.4.2
pandas==3.0.0
sentence-transformers==5.2.2
scikit-learn==1.8.0
pyarrow==23.0.0